+++ lib/src/constants/options_constant_keys.dart
@@ -2,5 +2,5 @@ abstract class OptionsConstantKey {
   static const String optionName = 'optionName';
   static const String optionIcon = 'optionIcon';
   static const String optionType = 'optionType';
-  static const String onSelectOption = 'onSelectOption';
+  static const String onSelectType = 'onSelectType';
 }